factor out key suppression into hook, suppress emission of characters on emacs keybindings
